requirements.txt
fs~=2.4.16

future~=0.18.2
Pillow~=9.3.0
psutil~=5.9.4
watchdog~=2.1.9
numpy~=1.23.4